+++ test/test_helper.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 require_relative '../lib/kanayago'
+require_relative '../lib/kanayago/cli'
 
 require 'minitest/autorun'